Fix the issue of logging in every time the home page is visited  users should only need to log in once at the start of the app
